Moves things around for session
