+++ lib/src/action/set_action.dart
@@ -1,4 +1,4 @@
-import '../data_store.dart';
+import '../store/mem_data_store.dart';
 import 'action_patch.dart';
 import 'action_result.dart';
 import 'base_action.dart';
