+++ lib/src/store/mem_data_store.dart
@@ -1,6 +1,6 @@
-import 'listener/handler.dart';
-import 'model/data_store.dart';
-import 'model/data_store_options.dart';
+import '../listener/handler.dart';
+import 'data_store.dart';
+import '../model/data_store_options.dart';
 
 /// A simple in-memory implementation of [BaseDataStore] using [Map].
 class KiwiWatermelonDataStore implements BaseDataStore {
